move icon images to icons folder
